refactor: Separate entity packages
