+++ src/main/java/com/paymentteamproject/domain/product/entity/Product.java
@@ -1,4 +1,4 @@
-package com.paymentteamproject.domain;
+package com.paymentteamproject.domain.product.entity;
 
 import com.paymentteamproject.common.entity.BaseEntity;
 import jakarta.persistence.*;
