+++ src/main/java/com/paymentteamproject/domain/product/entity/ProductCategory.java
@@ -1,4 +1,4 @@
-package com.paymentteamproject.domain;
+package com.paymentteamproject.domain.product.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
