+++ src/main/java/com/paymentteamproject/domain/webhook/entity/PaymentWebhookPaymentStatus.java
@@ -1,4 +1,4 @@
-package com.paymentteamproject.domain.webhook;
+package com.paymentteamproject.domain.webhook.entity;
 
 public enum PaymentWebhookPaymentStatus {
     READY,
